test(dev): fold the Workflow root cases into the app loader suite
`tests/integration/agent_loader/` is about one narrow thing -- that `__dirname`,
`__filename` and `import.meta.url` survive compilation -- while everything about
*what a loaded entrypoint may export* already lives in
`tests/integration/app_loader/`: an `app.ts`, an `app.js`, a default export, and
a `discovery` project that mixes App directories, agent directories and
standalone files. A bare `Workflow` root is another entry in that same table, not
a suite of its own, so a reader comparing the accepted setups sees them together
and one `npm install` per fixture covers them all.

The cases are unchanged in substance, only rehomed:

- `app_workflow/` joins the CLI matrix, so `adk run` on an entrypoint exporting a
  graph is exercised next to the App entrypoints. Its answer node greets, which
  is the assertion the matrix already shares.
- `discovery/service_graph/` makes the graph a directory entrypoint alongside
  `service_alpha` (App) and `service_beta` (agent): discovery lists it, the
  compiled root adapts into a `WorkflowAgent`, it runs through a real `Runner`,
  and `loadApp()` synthesizes an App around it.
- `discovery/lone_node.ts` keeps the other half honest -- a node that is not a
  `Workflow` is still not a root, so it stays out of `listAgents()` and loading
  it by hand still throws.

Real compilation stays the point: nothing is mocked, so the `Workflow` the loader
inspects still comes from the second, bundled copy of `@google/adk` that the
brand check exists for.
