+++ tests/integration/app_loader/app_loader_test.ts
@@ -4,21 +4,54 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {App, isApp, isBaseAgent} from '@google/adk';
+import {
+  App,
+  BaseAgent,
+  Event,
+  InMemorySessionService,
+  isApp,
+  isBaseAgent,
+  isGraphWorkflowAgent,
+  Runner,
+} from '@google/adk';
 import {exec, spawn} from 'node:child_process';
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
 import {promisify} from 'node:util';
 import {afterAll, beforeAll, describe, expect, it} from 'vitest';
-import {AgentLoader} from '../../../dev/src/utils/agent_loader.js';
+import {AgentFile, AgentLoader} from '../../../dev/src/utils/agent_loader.js';
 import {sendInput} from '../test_case_utils.js';
 
 const execAsync = promisify(exec);
 const dirname = process.cwd();
 const TEST_EXECUTION_TIMEOUT = 60000;
 
+async function runToCompletion(agent: BaseAgent): Promise<Event[]> {
+  const sessionService = new InMemorySessionService();
+  const session = await sessionService.createSession({
+    appName: 'app_loader',
+    userId: 'test_user',
+  });
+  const runner = new Runner({
+    appName: 'app_loader',
+    agent,
+    sessionService,
+  });
+
+  const events: Event[] = [];
+  for await (const event of runner.runAsync({
+    userId: 'test_user',
+    sessionId: session.id,
+    newMessage: {role: 'user', parts: [{text: 'Hello Graph'}]},
+  })) {
+    events.push(event);
+  }
+
+  return events;
+}
+
 describe('App loader CLI integration', () => {
-  describe.each(['app_ts', 'app_js', 'app_default'])(
+  describe.each(['app_ts', 'app_js', 'app_default', 'app_workflow'])(
     'App entrypoint with %s',
     (testCaseName: string) => {
       const projectPath = path.join(
@@ -88,11 +121,16 @@ describe('AgentLoader discovery and loading integration', () => {
       expect(apps).toContain('standalone_app');
 
       const agentsAndApps = await loader.listAgents();
-      expect(agentsAndApps).toHaveLength(4);
+      expect(agentsAndApps).toHaveLength(5);
       expect(agentsAndApps).toContain('service_alpha');
       expect(agentsAndApps).toContain('service_beta');
+      // Before a Workflow could be a root this was not an error, it was a
+      // silence: the file exported nothing matching `isBaseAgent`, so the
+      // directory simply did not show up.
+      expect(agentsAndApps).toContain('service_graph');
       expect(agentsAndApps).toContain('standalone_agent');
       expect(agentsAndApps).toContain('standalone_app');
+      expect(await loader.listLoadFailures()).toEqual([]);
     },
     TEST_EXECUTION_TIMEOUT,
   );
@@ -123,6 +161,72 @@ describe('AgentLoader discovery and loading integration', () => {
       const synthApp = await agentFile.loadApp();
       expect(isApp(synthApp)).toBe(true);
       expect(synthApp.rootAgent.name).toBe('beta_agent');
+    },
+    TEST_EXECUTION_TIMEOUT,
+  );
+
+  /**
+   * Real compilation is the point of these cases. The loader's unit tests mock
+   * `esbuild.build` into a file copy, which leaves the interesting part of the
+   * bundle untested: `packages: 'bundle'` inlines `@google/adk` into the
+   * compiled agent, so the `Workflow` the loader inspects comes from a *second*
+   * copy of the library. That is the case the
+   * `Symbol.for('google.adk.workflow.workflow')` brand exists for, and the case
+   * an `instanceof` check would silently fail.
+   */
+  it(
+    'should adapt a compiled Workflow export into a runnable root agent',
+    async () => {
+      const agentFile = await loader.getAgentFile('service_graph');
+      const rootAgent = await agentFile.loadAgent();
+
+      expect(isBaseAgent(rootAgent)).toBe(true);
+      // Still a WorkflowAgent, which is what the dev server's graph renderer
+      // and the a2a card match on.
+      expect(isGraphWorkflowAgent(rootAgent)).toBe(true);
+      expect(rootAgent.name).toBe('graph_workflow');
+      expect(rootAgent.description).toBe(
+        'Normalizes the question, then answers it.',
+      );
+    },
+    TEST_EXECUTION_TIMEOUT,
+  );
+
+  it(
+    'should run a loaded Workflow root through a real Runner',
+    async () => {
+      const agentFile = await loader.getAgentFile('service_graph');
+      const events = await runToCompletion(await agentFile.loadAgent());
+
+      // Both nodes ran, in order, on the user's message.
+      expect(
+        events.map((event) => event.output).filter((o) => o !== undefined),
+      ).toEqual(['hello graph', 'graph handled: hello graph']);
+    },
+    TEST_EXECUTION_TIMEOUT,
+  );
+
+  it(
+    'should synthesize an App around a Workflow root',
+    async () => {
+      const appFile = await loader.getAppFile('service_graph');
+      const app = await appFile.loadApp();
+
+      expect(isApp(app)).toBe(true);
+      expect(app.rootAgent.name).toBe('graph_workflow');
+    },
+    TEST_EXECUTION_TIMEOUT,
+  );
+
+  it(
+    'should still refuse a node that is not a Workflow',
+    async () => {
+      expect(await loader.listAgents()).not.toContain('lone_node');
+
+      const agentFile = new AgentFile(path.join(projectPath, 'lone_node.ts'));
+      await expect(agentFile.load()).rejects.toThrow(
+        /No @google\/adk BaseAgent or Workflow instance found/,
+      );
     },
     TEST_EXECUTION_TIMEOUT,
   );
